ppd-ipp.c: Use correct number of items to prevent buffer overflow
The array members are strings, so we can get correct number of times by
dividing the array size by char* size.

debug.c: Protect against possible format string attack

ppd-collection.cxx: Use intmax_t for printing time_t var

+++ ppd/ppd-collection.cxx
@@ -848,9 +848,9 @@ ppdCollectionDumpCache(const char *filename,	// I - Filename
   for (ppd = (ppd_info_t *)cupsArrayGetFirst(ppdlist.PPDsByName);
        ppd;
        ppd = (ppd_info_t *)cupsArrayGetNext(ppdlist.PPDsByName))
-    printf("%d,%ld,%d,%d,\"%s\",\"%s\",\"%s\",\"%s\",\"%s\",\"%s\",\"%s\","
+    printf("%jd,%ld,%d,%d,\"%s\",\"%s\",\"%s\",\"%s\",\"%s\",\"%s\",\"%s\","
            "\"%s\",\"%s\"\n",
-           (int)ppd->record.mtime, (long)ppd->record.size,
+           (intmax_t)ppd->record.mtime, (long)ppd->record.size,
 	   ppd->record.model_number, ppd->record.type, ppd->record.filename,
 	   ppd->record.name, ppd->record.languages[0], ppd->record.products[0],
 	   ppd->record.psversions[0], ppd->record.make,
